refactor: move CUP utilities to src/integration/cup/utility and update imports
