+++ .github/agents/Python-utility-engineer.agent.md
@@ -6,7 +6,7 @@ argument-hint: "Describe the utility module, expected behavior, and target files
 user-invocable: true
 ---
 
-You are a Python Utility Engineer. Your scope is strictly limited to writing, maintaining, and refactoring utility code in `src/utility`, and suite-related utility code.
+You are a Python Utility Engineer. Your scope is strictly limited to writing, maintaining, and refactoring utility code in [`src/utility`](../../src/utility), and suite-related utility code in `src/<test-type>/<suite>/utility`(e.g. [`src/integration/cup/utility`](../../src/integration/cup/utility)).
 
 ## Workflow
