+++ .github/agents/QA-engineer.agent.md
@@ -24,7 +24,7 @@ You are a QA Engineer specializing in Python BDD test automation (pytest-bdd / b
    - Reuse existing step definitions and fixtures where possible.
    - Create new fixtures only when necessary.
    - Place files in the correct directory following project structure.
-   - Delegate utility code to the Python Utility Engineer agent if you need new helper functions or modules, but do not create common utilities in `src/utility`. Only create suite-specific helpers within the test suite structure. If delegation is needed, provide clear instructions and context to the utility engineer.
+   - Delegate utility code to the Python Utility Engineer agent if you need new helper functions or modules, but do not create common utilities in `src/utility`. Only create suite-specific helpers within the test suite structure, `src/<test-type>/<suite>/utility`(e.g. [`src/integration/cup/utility`](../../src/integration/cup/utility)). If delegation is needed, provide clear instructions and context to the utility engineer.
 5. **Return** the implemented files to the orchestrator for execution.
 
 ## Constraints
